\name{print.criterionRkh}
\alias{print.criterionRkh}
\title{
Print a Summary of a criterionRkh Object
}
\description{
This is a method for the function print for objects of the class \code{criterionRkh}. 
}

\usage{
\method{print}{criterionRkh}(x, \dots)
}

\arguments{
  \item{x}{
An object of class  \code{criterionRkh} generated by the function \code{\link{criterionRkh}}. 
}
  \item{\dots}{
Further arguments passed to or from other methods.
}
}

\value{
A summary of the \code{criterionRkh} Object is returned.
}


\author{
Benoit Liquet, <benoit.liquet@isped.u-bordeaux2.fr> and Jerome Saracco <jerome.saracco@math.u-bordeaux1.fr>
}


\seealso{
\code{\link{criterionRkh}}
}

\keyword{ print }

