\name{print.edr}
\alias{print.edr}

\title{
Print a Summary of an edr Object
}

\description{
This is a method for the function print for objects of the class \code{edr}.
}

\usage{
\method{print}{edr}(x, \dots)
}

\arguments{
  \item{x}{An object of class  \code{edr} generated by the function \code{\link{edr}}. 
}
  \item{\dots}{
Further arguments to be passed to or from other methods. They are ignored in this function
}
}

\value{
A summary of the \code{edr} Object is returned.
}


\author{
Benoît Liquet, <benoit.liquet@isped.u-bordeaux2.fr> and Jérôme Saracco <jerome.saracco@math.u-bordeaux1.fr>
}


\seealso{
\code{\link{edr}}
}

\keyword{ print }

