\name{plot.edrSelec}
\alias{plot.edrSelec}
\alias{print.edrSelec}
\alias{summary.edrSelec}

\title{
Displaying methods for the class 'edrSelec' 
}

\description{
Methods to present the results of a variable selection procedure based on slice inverse regression.
}

\usage{
\method{print}{edrSelec}(x, ...)
\method{summary}{edrSelec}(object, nVar=5, ...)
\method{plot}{edrSelec}(x, nVar=25, ...)
}

\arguments{
  \item{x}{An object of class  \code{edrSelec} generated by the function \code{\link{edrSelec}}. 
}
\item{object}{An object of class  \code{edrSelec} generated by the function \code{\link{edrSelec}}. 
}
  \item{nVar}{
	If \code{nVar!=NULL}, the method will only show the \code{nVar} most important variables with respect to the criterion computed by the function \code{\link{edrSelec}}. Not implemented for the \code{summary} method when executed with a "SR-SIR" procedure. 
}

	\item{...}{Other unused parameters.}
}

\value{
These methods display (with a text or a graphic) the score of each considered variable. The user should keep the variables with a high score.
}


\author{
Raphaël Coudret <rcoudret@gmail.com>, Benoît Liquet <benoit.liquet@isped.u-bordeaux2.fr> and Jérôme Saracco <jerome.saracco@math.u-bordeaux1.fr>
}


\seealso{
\code{\link{edrSelec}}, \code{\link{edrUnderdet}}, \code{\link{edr}}
}

\examples{
## see the example of the function edrSelec.
}

\keyword{ print }
\keyword{ hplot }
\keyword{ methods }

