\name{summary.criterionRkh}
\alias{summary.criterionRkh}

\title{
Print a Summary of a criterionRkh Object
}
\description{
This is a method for the function summary for objects of the class \code{criterionRkh}.
}
\usage{
\method{summary}{criterionRkh}(object, \dots)
}

\arguments{
  \item{object}{
An object of class  \code{criterionRkh} generated by the function \code{\link{criterionRkh}}. 
}
  \item{\dots}{
Further arguments passed to or from other methods.
}
}

\value{
A summary of the \code{criterionRkh} Object is returned.
}
\author{
Benoît Liquet, <benoit.liquet@isped.u-bordeaux2.fr> and Jérôme Saracco <jerome.saracco@math.u-bordeaux1.fr>
}

\seealso{
 \code{\link{criterionRkh}}
}

\keyword{ regression}
