\name{summary.edr}
\alias{summary.edr}

\title{
Print a Summary of a edr Object
}
\description{
This is a method for the function summary for objects of the class \code{edr}.
}
\usage{
\method{summary}{edr}(object, \dots)
}

\arguments{
  \item{object}{
An object of class  \code{edr} generated by the function \code{\link{edr}}. 
}
  \item{\dots}{
further arguments passed to or from other methods.
}
}


\value{
A summary of the \code{edr} Object is returned.
}
\author{
Benoit Liquet, <benoit.liquet@isped.u-bordeaux2.fr> and Jerome Saracco <jerome.saracco@math.u-bordeaux1.fr>
}

\seealso{
 \code{\link{edr}}
}

\keyword{ regression}